sync RMF to 5034df6feff9073f7d5746e63e975bebc57173b6
git-svn-id: https://svn.salilab.org/imp/trunk@15839 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/rmf/dependency/RMF/src/backend/avro/AvroSharedData.h
@@ -85,7 +85,8 @@ namespace RMF {
         std::string keyname= P::get_key_name(k.get_id());
         typename std::map<std::string, int>::const_iterator it
           =index.find(keyname);
-        if (it==index.end() || it->second >= data.size()) {
+        if (it==index.end()
+            || it->second >= static_cast<int>(data.size())) {
           return TypeTraits::get_null_value();
         } else {
           typename TypeTraits::Type ret;
@@ -110,7 +111,7 @@ namespace RMF {
         } else {
           index_value=it->second;
         }
-        if (data.size() <= index_value) {
+        if (static_cast<int>(data.size()) <= index_value) {
           typename TypeTraits::AvroType null;
           avro_assign(null, TypeTraits::get_null_value());
           data.resize(index_value+1, null);
